# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "An API for users to register/login and view different products.\n                    Registered users can add products to their cart and checkout to create \n                    their order. Registered users with admin role can add, edit and delete the products."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Get all products. Admins receive all products, while other users or guests get only available products.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched list of products (Only non-deleted products if not admin)",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Products found"
                    },
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Create a new product. Admins only.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Product data to create",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Example product"
                },
                "description": {
                  "type": "string",
                  "example": "Product description"
                },
                "unitprice": {
                  "type": "number",
                  "example": 499
                },
                "imgurl": {
                  "type": "string",
                  "example": "http://images.test.png"
                },
                "quantity": {
                  "type": "number",
                  "example": 50
                },
                "brand_id": {
                  "type": "number",
                  "example": 1
                },
                "category_id": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Product created successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 201
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Product created"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad request — missing required fields.",
            "schema": {
              "$ref": "#components/schemas/MissingParameters"
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "403": {
            "description": "User is not admin",
            "schema": {
              "$ref": "#components/schemas/NotAdmin"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Get all product by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched product",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Product found"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "404": {
            "description": "Product not found",
            "schema": {
              "$ref": "#components/schemas/ResourceNotFound"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Update a product by ID. Admins only",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the product to update",
            "example": 1
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Product fields to update",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Updated product name"
                },
                "description": {
                  "type": "string",
                  "example": "Updated description"
                },
                "unitprice": {
                  "type": "number",
                  "example": 999
                },
                "imgurl": {
                  "type": "string",
                  "example": "http://new.image.url.png"
                },
                "quantity": {
                  "type": "number",
                  "example": 100
                },
                "brand": {
                  "type": "string",
                  "example": "Updated brand"
                },
                "category": {
                  "type": "string",
                  "example": "Updated category"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Product updated"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad request — missing required fields.",
            "schema": {
              "$ref": "#components/schemas/MissingParameters"
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "403": {
            "description": "User is not admin",
            "schema": {
              "$ref": "#components/schemas/NotAdmin"
            }
          },
          "404": {
            "description": "Product not found",
            "schema": {
              "$ref": "#components/schemas/ResourceNotFound"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Delete a product by ID. Admins only",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the product to update",
            "example": 1
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Product deleted"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad request — missing required fields.",
            "schema": {
              "$ref": "#components/schemas/MissingParameters"
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "403": {
            "description": "User is not admin",
            "schema": {
              "$ref": "#components/schemas/NotAdmin"
            }
          },
          "404": {
            "description": "Product not found",
            "schema": {
              "$ref": "#components/schemas/ResourceNotFound"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Example name"
          },
          "description": {
            "type": "string",
            "example": "Example description"
          },
          "unitprice": {
            "type": "integer",
            "example": 649
          },
          "imgurl": {
            "type": "string",
            "example": "http://images.test.png"
          },
          "quantity": {
            "type": "integer",
            "example": 10
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "brand_id": {
            "type": "integer",
            "example": 1
          },
          "category_id": {
            "type": "integer",
            "example": 1
          },
          "brand": {
            "type": "string",
            "example": "Example brand"
          },
          "category": {
            "type": "string",
            "example": "Example category"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-02T09:25:53Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-02T09:25:53Z"
          }
        }
      },
...
    }
  }
}